Move ManyToMany test fixtures into their test classes
