--- test/EFCore.SqlServer.FunctionalTests/Query/ManyToManyQuerySqlServerFixture.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.EntityFrameworkCore.TestUtilities;
-
-namespace Microsoft.EntityFrameworkCore.Query
-{
-    public class ManyToManyQuerySqlServerFixture : ManyToManyQueryRelationalFixture
-    {
-        protected override ITestStoreFactory TestStoreFactory => SqlServerTestStoreFactory.Instance;
-    }
-}
