+++ test/EFCore.Sqlite.FunctionalTests/Query/ManyToManyQuerySqliteTest.cs
@@ -1,13 +1,20 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.EntityFrameworkCore.TestUtilities;
+
 namespace Microsoft.EntityFrameworkCore.Query
 {
-    public class ManyToManyQuerySqliteTest : ManyToManyQueryTestBase<ManyToManyQuerySqliteFixture>
+    public class ManyToManyQuerySqliteTest : ManyToManyQueryTestBase<ManyToManyQuerySqliteTest.ManyToManyQuerySqliteFixture>
     {
         public ManyToManyQuerySqliteTest(ManyToManyQuerySqliteFixture fixture)
             : base(fixture)
         {
+        }
+
+        public class ManyToManyQuerySqliteFixture : ManyToManyQueryRelationalFixture
+        {
+            protected override ITestStoreFactory TestStoreFactory => SqliteTestStoreFactory.Instance;
         }
     }
 }
